requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
debugpy==1.8.1
Django==5.0.2
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mysqlclient==2.2.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
setuptools==68.2.2
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.41.2